+++ packages/common-styles/src/animations.less
