+++ packages/component-button/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "skeoh-ui-component-flex",
+  "name": "@skeoh-ui/component-button",
   "version": "0.0.0",
-  "description": "Flexbox utility components",
+  "description": "Generic and specific buttons",
   "typings": "./src/index.tsx",
   "main": "./dist/index.js",
   "scripts": {
@@ -19,6 +19,8 @@
   },
   "homepage": "https://github.com/j-/skeoh-ui#readme",
   "dependencies": {
-    "skeoh-ui-common-styles": "*"
+    "@skeoh-ui/common-styles": "*",
+    "@skeoh-ui/component-fill": "*",
+    "@skeoh-ui/component-flex": "*"
   }
 }
