+++ packages/component-button/src/styles.less
@@ -1,6 +1,6 @@
-@import "../node_modules/skeoh-ui-common-styles/src/styles";
-@import "../node_modules/skeoh-ui-component-flex/src/styles";
-@import "../node_modules/skeoh-ui-component-fill/src/styles";
+@import "../node_modules/@skeoh-ui/common-styles/src/styles";
+@import "../node_modules/@skeoh-ui/component-flex/src/styles";
+@import "../node_modules/@skeoh-ui/component-fill/src/styles";
 
 .skeoh-ui-button {
 	padding: 0.25em;
