+++ packages/component-fill/src/styles.less
@@ -1,4 +1,4 @@
-@import "../node_modules/skeoh-ui-common-styles/src/styles";
+@import "../node_modules/@skeoh-ui/common-styles/src/styles";
 
 .skeoh-ui-fill,
 .skeoh-ui-fill-horizontal {
