+++ packages/component-flex/src/styles.less
@@ -1,0 +1,5 @@
+@import "../node_modules/@skeoh-ui/common-styles/src/styles";
+
+.skeoh-ui-flex-container {
+	display: flex;
+}
