+++ packages/component-input/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "skeoh-ui-component-button",
+  "name": "@skeoh-ui/component-input",
   "version": "0.0.0",
-  "description": "Generic and specific buttons",
+  "description": "Form control components",
   "typings": "./src/index.tsx",
   "main": "./dist/index.js",
   "scripts": {
@@ -19,8 +19,6 @@
   },
   "homepage": "https://github.com/j-/skeoh-ui#readme",
   "dependencies": {
-    "skeoh-ui-common-styles": "*",
-    "skeoh-ui-component-fill": "*",
-    "skeoh-ui-component-flex": "*"
+    "@skeoh-ui/common-styles": "*"
   }
 }
