+++ packages/component-input/src/styles.less
@@ -1,4 +1,4 @@
-@import "../node_modules/skeoh-ui-common-styles/src/styles";
+@import "../node_modules/@skeoh-ui/common-styles/src/styles";
 
 .skeoh-ui-input {
 	padding: 0.25em;
