+++ packages/guide/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "skeoh-ui-component-fill",
+  "name": "@skeoh-ui/guide",
   "version": "0.0.0",
-  "description": "Space filling utility components",
-  "typings": "./src/index.tsx",
-  "main": "./dist/index.js",
+  "description": "UI kit demos and style guide",
+  "main": "index.js",
   "scripts": {
-    "build": "tsc && lessc src/styles.less dist/styles.css",
+    "build": "cross-env NODE_ENV=production webpack -p",
+    "start": "webpack-dev-server --inline --content-base dist/",
     "test": "echo \"Error: no test specified\""
   },
   "repository": {
@@ -19,6 +19,9 @@
   },
   "homepage": "https://github.com/j-/skeoh-ui#readme",
   "dependencies": {
-    "skeoh-ui-common-styles": "*"
+    "@skeoh-ui/component-button": "*",
+    "@skeoh-ui/component-fill": "*",
+    "@skeoh-ui/component-flex": "*",
+    "@skeoh-ui/component-input": "*"
   }
 }
