+++ packages/guide/src/components/Demo.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as CodeMirror from 'react-codemirror';
-import { Submit } from 'skeoh-ui-component-button/src';
-import { FillHorizontal } from 'skeoh-ui-component-fill/src';
+import { Submit } from '@skeoh-ui/component-button/src';
+import { FillHorizontal } from '@skeoh-ui/component-fill/src';
 import * as ts from 'typescript';
 import bootstrap from '../safe-imports';
 
@@ -34,11 +34,11 @@ import * as ReactDOM from 'react-dom';
 import {
   Button,
   ButtonGroup,
-} from 'skeoh-ui-component-button';
+} from '@skeoh-ui/component-button';
 
 import {
   FillHorizontal,
-} from 'skeoh-ui-component-fill';
+} from '@skeoh-ui/component-fill';
 
 ReactDOM.render(
   <div>
