+++ packages/guide/src/styles/index.less
@@ -1,7 +1,7 @@
 @import (inline) "../../../../node_modules/codemirror/lib/codemirror.css";
 
-@import "../../node_modules/skeoh-ui-component-button/src/styles";
-@import "../../node_modules/skeoh-ui-component-input/src/styles";
+@import "../../node_modules/@skeoh-ui/component-button/src/styles";
+@import "../../node_modules/@skeoh-ui/component-input/src/styles";
 
 body {
 	margin: 1em;
