--- packages/skeoh-ui-common-styles/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "skeoh-ui-common-styles",
-  "version": "0.0.0",
-  "description": "Styles to be shared between components",
-  "scripts": {
-    "build": "lessc src/styles.less dist/styles.css",
-    "test": "echo \"Error: no test specified\""
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/j-/skeoh-ui.git"
-  },
-  "author": "Jamie Hoeks <j@skeoh.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/j-/skeoh-ui/issues"
-  },
-  "homepage": "https://github.com/j-/skeoh-ui#readme"
-}
