--- packages/skeoh-ui-component-flex/src/styles.less
@@ -1,5 +1,0 @@
-@import "../node_modules/skeoh-ui-common-styles/src/styles";
-
-.skeoh-ui-flex-container {
-	display: flex;
-}
